Create board file
